♻️ Split account page into section card components
Each card owns its domain logic via composables; the page keeps only
the login watcher, deep-link handling, and shared layout. Stripe
Connect status moves to useState so the page watcher and creator card
share it, and the adult-content modal emits confirm for the opener to
apply (per-instance logged-out setting state).
